mv sendConfigChangeRequest to lib
cleanup fedi checks
